More consistent I/O API.

+++ src/core/vm/modules/source.c
@@ -92,14 +92,14 @@ static size_t _handle_read(void *user_data, void *buffer, size_t bytes_to_read)
     return FS_read(handle, buffer, bytes_to_read);
 }
 
-static bool _handle_seek(void *user_data, int offset, int whence)
+static bool _handle_seek(void *user_data, long offset, int whence)
 {
     File_System_Handle_t *handle = (File_System_Handle_t *)user_data;
     FS_seek(handle, offset, whence);
     return true;
 }
 
-static int _handle_tell(void *user_data)
+static long _handle_tell(void *user_data)
 {
     File_System_Handle_t *handle = (File_System_Handle_t *)user_data;
     return FS_tell(handle);
