+++ src/libs/fs/fs.c
@@ -186,7 +186,7 @@ void FS_seek(File_System_Handle_t *handle, long offset, int whence)
     ((Handle_t *)handle)->vtable.seek(handle, offset, whence);
 }
 
-int FS_tell(File_System_Handle_t *handle)
+long FS_tell(File_System_Handle_t *handle)
 {
     return ((Handle_t *)handle)->vtable.tell(handle);
 }
