+++ src/libs/fs/pak.c
@@ -94,7 +94,7 @@ static void _pak_handle_dtor(File_System_Handle_t *handle);
 static size_t _pak_handle_size(File_System_Handle_t *handle);
 static size_t _pak_handle_read(File_System_Handle_t *handle, void *buffer, size_t bytes_requested);
 static void _pak_handle_seek(File_System_Handle_t *handle, long offset, int whence);
-static int _pak_handle_tell(File_System_Handle_t *handle);
+static long _pak_handle_tell(File_System_Handle_t *handle);
 static bool _pak_handle_eof(File_System_Handle_t *handle);
 
 bool pak_is_valid(const char *path)
@@ -400,7 +400,7 @@ static void _pak_handle_seek(File_System_Handle_t *handle, long offset, int when
     Log_write(LOG_LEVELS_DEBUG, LOG_CONTEXT, "%d bytes seeked w/ mode %d for handle %p", offset, whence, handle);
 }
 
-static int _pak_handle_tell(File_System_Handle_t *handle)
+static long _pak_handle_tell(File_System_Handle_t *handle)
 {
     Pak_Handle_t *pak_handle = (Pak_Handle_t *)handle;
